Refactor training modules and update import paths for better organization
