+++ src/training/validator.py
@@ -3,7 +3,7 @@
 import time
 from typing import Any
 
-from src.training.training_validation_config import (
+from src.training.steps.data_preparation_components.training_validation_config import (
     VALIDATION_FUNCTIONS,
     can_proceed_to_step,
     get_progression_rules,
